+++ src/DataFixtures/AppFixtures.php
@@ -5,30 +5,62 @@ namespace App\DataFixtures;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
 use App\Entity\Announcement;
-use App\DataFixtures\UserFixtures;
+use App\Entity\User;
 use Doctrine\Common\DataFixtures\DependentFixtureInterface;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
+
 
 
 class AppFixtures extends Fixture
 {
+  private $passwordHasher;
+
+  public function __construct(UserPasswordHasherInterface $passwordHasher)
+  {
+    $this->passwordHasher = $passwordHasher;
+  }
+
   public function load(ObjectManager $manager)
   {
 
+    $user = new User();
+    $user->setUsername('david');
+    $user->setPassword($this->passwordHasher->hashPassword(
+        $user,
+        '12345'
+    ));
+    $manager->persist($user);
+
+    // one announcement set to a past date, one to the current date (constant) and one for a future date
     $announcement = new Announcement();
     $announcement_date = new \DateTime('2021-07-14');
     $announcement->setSubject('autoSubject');
     $announcement->setAuthor('autoAuthor');
-    $announcement->setUser($this->getReference(UserFixtures::ADMIN_USER_REFERENCE));
+    $announcement->setUser($user);
     $announcement->setDate($announcement_date);
     $announcement->setText('autoText');
+    $manager->persist($announcement);
 
-  }
+    $announcement = new Announcement();
+    $announcement_date = new \DateTime('now');
+    $announcement->setSubject('autoSubject');
+    $announcement->setAuthor('autoAuthor');
+    $announcement->setUser($user);
+    $announcement->setDate($announcement_date);
+    $announcement->setText('autoText');
+    $manager->persist($announcement);
 
-  public function getDependencies()
-  {
-    return [
-      UserFixtures::class,
-    ];
+    $announcement = new Announcement();
+    $announcement_date = new \DateTime('3021-07-14');
+    $announcement->setSubject('autoSubject');
+    $announcement->setAuthor('autoAuthor');
+    $announcement->setUser($user);
+    $announcement->setDate($announcement_date);
+    $announcement->setText('autoText');
+    $manager->persist($announcement);
+
+    $manager->flush();
+
   }
 
 }
